Added solution - LeetHub

+++ Count pairs in array divisible by K - GFG/count-pairs-in-array-divisible-by-k.cpp
@@ -14,7 +14,7 @@ class Solution
     long long countKdivPairs(int A[], int n, int K)
     {
         //code here
-        unordered_map<long long  , long long > mp;
+        unordered_map<int , int > mp;
         long long  ans=0;
         for(int i=0 ; i<n ; i++ ){
